app/fix: changed accumulatedCoffeeBuys and accumulatedExpenses to Double

+++ coffee-order-application/app/src/main/java/com/feup/coffee_order_application/models/Data.kt
@@ -2,8 +2,8 @@ package com.feup.coffee_order_application.models
 data class Data(
     val __v: Int,
     val _id: String,
-    val accumulatedCoffeeBuys: Int,
-    val accumulatedExpenses: Int,
+    val accumulatedCoffeeBuys: Double,
+    val accumulatedExpenses: Double,
     val email: String,
     val name: String,
     val nif: String
